+++ src/Controller/Admin/PointOfInterestController.php
@@ -33,6 +33,8 @@ class PointOfInterestController extends AbstractCrudController
         ->setHelp(new TranslatableMessage('The latitude of the interest point.', [], 'admin'));
         yield TextField::new('longitude')->setRequired(true)
         ->setHelp(new TranslatableMessage('The longitude of the interest point.', [], 'admin'));
+        yield NumberField::new('proximityToUnlock')
+        ->setHelp(new TranslatableMessage('The proximity that allows unlocking this point of interest (in m).', [], 'admin'));
 
         $context = $this->getContext();
         if (in_array($pageName, [Crud::PAGE_NEW, Crud::PAGE_EDIT], true) && null !== $context) {